Modified converter to keep epoch time as number, not string
Debated between making a new converter for the epoch times or using
the existing one. After implementing a separate one, I felt it
would be easier to have a single entry point for date converters
and alternate using the existing Format enumeration. Sharing a
single class does reduce the usefulness of the generic type, but
it's all used internally to the view creation anyways, so didn't
feel it was a big loss.